refactor(boarding-pass): split 708-line widget into 3 part files
708→157/225/331 lines. Extract _HeaderStrip+_EventInfoSection+_FieldLabel
into boarding_pass_card_info.dart, and _PerforationLine+_QRStubSection
+related painters/badges into boarding_pass_card_qr.dart. Uses Dart
part-of to share library scope — no API change, no import changes needed
in callers.

Constraint: Private widget classes (_Foo) must stay in same Dart library
Confidence: high
Scope-risk: narrow
